now gets vod links
